Extracts article dedupe util and adds typewriter caret buffer
Centralizes article deduplication into a shared utility to remove duplicated logic and ensure consistent deduping across the app.

Introduces a small caret-start buffer for the typewriter headline animation and applies it to the animator to prevent visual overlap between typing and subsequent reveals, making timing easier to tune.
